requirements.txt
Flask
openai
chromadb
PyMuPDF
python-dotenv